Add custom CA support for azure (#4974)
* Add custom CA support for azure


* Update CHANGELOG.md


* have a shared tlsconfig.go
